requirements.txt
boto==2.38.0
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==1.9.13
django-admin-hstore-widget==1.0.1
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.4.0
django-import-export==1.1.0
django-pipeline==1.6.14
django-sass-processor==0.7
django-test-without-migrations==0.4
djangorestframework==3.6.4
ecdsa==0.13
elastic-apm==3.0.1
et-xmlfile==1.0.1
futures==3.2.0
GDAL==1.10
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
libsass==0.14.5
Markdown==3.0.1
MarkupSafe==1.0
openpyxl==2.5.9
paramiko==2.0.9
Pillow==5.3.0
psycopg2==2.7.5
pycrypto==2.6.1
python-logstash==0.4.6
rcssmin==1.0.6
requests==2.20.0
rjsmin==1.0.12
Shapely==1.6.4.post2
six==1.11.0
SQLAlchemy==1.2.11
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
git+https://github.com/OpenUpSA/wazimap.git@releases#egg=wazimap[gdal]
whitenoise==4.0
django-storages==1.1.8
